// admin/src/api/generated/models/ExternalApiKeyCreatedResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * deploy-go-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ExternalApiKeyCreatedResponse
 */
export interface ExternalApiKeyCreatedResponse {
    /**
     *
     * @type {Array<string>}
     * @memberof ExternalApiKeyCreatedResponse
     */
    applicationIds: Array<string>;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    createdAt: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    expiresAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    status: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeyCreatedResponse
     */
    token: string;
}

/**
 * Check if a given object implements the ExternalApiKeyCreatedResponse interface.
 */
export function instanceOfExternalApiKeyCreatedResponse(value: object): value is ExternalApiKeyCreatedResponse {
    if ((!('applicationIds' in (value as Record<string, any>)) && !('application_ids' in (value as Record<string, any>))) || ((value as Record<string, any>)['applicationIds'] === undefined && (value as Record<string, any>)['application_ids'] === undefined)) return false;
    if ((!('createdAt' in (value as Record<string, any>)) && !('created_at' in (value as Record<string, any>))) || ((value as Record<string, any>)['createdAt'] === undefined && (value as Record<string, any>)['created_at'] === undefined)) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('token' in value) || value['token'] === undefined) return false;
    return true;
}

export function ExternalApiKeyCreatedResponseFromJSON(json: any): ExternalApiKeyCreatedResponse {
    return ExternalApiKeyCreatedResponseFromJSONTyped(json, false);
}

export function ExternalApiKeyCreatedResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExternalApiKeyCreatedResponse {
    if (json == null) {
        return json;
    }
    return {

        'applicationIds': json['application_ids'],
        'createdAt': json['created_at'],
        'expiresAt': json['expires_at'] === undefined ? undefined : json['expires_at'] === null ? null : json['expires_at'],
        'id': json['id'],
        'name': json['name'],
        'status': json['status'],
        'token': json['token'],
    };
}

export function ExternalApiKeyCreatedResponseToJSON(json: any): ExternalApiKeyCreatedResponse {
    return ExternalApiKeyCreatedResponseToJSONTyped(json, false);
}

export function ExternalApiKeyCreatedResponseToJSONTyped(value?: ExternalApiKeyCreatedResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'application_ids': value['applicationIds'],
        'created_at': value['createdAt'],
        'expires_at': value['expiresAt'],
        'id': value['id'],
        'name': value['name'],
        'status': value['status'],
        'token': value['token'],
    };
}