// admin/src/api/generated/models/ExternalApiKeySummary.ts
export interface ExternalApiKeySummary {
    /**
     *
     * @type {Array<string>}
     * @memberof ExternalApiKeySummary
     */
    applicationIds: Array<string>;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    createdAt: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    expiresAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    lastUsedAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    status: string;
    /**
     *
     * @type {string}
     * @memberof ExternalApiKeySummary
     */
    updatedAt: string;
    /**
     *
     * @type {number}
     * @memberof ExternalApiKeySummary
     */
    version: number;
}